pyproject.toml
[build-system]
requires = ["setuptools>=45", "wheel"]
build-backend = "setuptools.build_meta"

[project]
name = "spatialkit"
version = "0.3.0"
description = "A spatial computing toolkit for 3D vision and robotics"
authors = [
    {name = "Sehyun Cha", email = "cshyundev@gmail.com"}
]
readme = "README.md"
license = {text = "MIT"}
requires-python = ">=3.12"
keywords = ["spatial-computing", "3d-vision", "geometry", "robotics", "pytorch", "computer-vision"]
classifiers = [
    "Development Status :: 3 - Alpha",
    "Intended Audience :: Developers",
    "Intended Audience :: Science/Research",
    "License :: OSI Approved :: MIT License",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.12",
    "Topic :: Scientific/Engineering :: Artificial Intelligence",
    "Topic :: Scientific/Engineering :: Image Processing",
]

dependencies = [
    "numpy>=1.24.0,<2.0.0",
    "opencv-python>=4.5.0",
    "torch>=2.0.0",
    "scipy>=1.11.0",
    "matplotlib>=3.7.0",
    "scikit-image>=0.21.0",
    "tifffile>=2023.7.0",
    "Pillow>=10.0.0",
    "PyYAML>=6.0",
]

[project.optional-dependencies]
markers = ["dt-apriltags>=3.1.7", "stag-python>=1.1.1"]
vis3d = ["open3d>=0.19.0"]
dev = ["pytest>=8.4.1", "pytest-cov", "absl-py>=2.3.1", "rich"]
all = ["spatialkit[markers,vis3d]"]

[project.urls]
Homepage = "https://github.com/cshyundev/spatialkit"
Repository = "https://github.com/cshyundev/spatialkit"
Issues = "https://github.com/cshyundev/spatialkit/issues"

[tool.setuptools]
package-dir = {"" = "src"}

[tool.setuptools.packages.find]
where = ["src"]

[tool.uv]
dev-dependencies = [
    "pytest>=6.0",
    "pytest-cov",
]
